+++ rio_tiler/constants.py
@@ -2,9 +2,14 @@
 
 import multiprocessing
 import os
+from typing import Tuple, Union
 
 import morecantile
 from rasterio.crs import CRS
+
+NumType = Union[float, int]
+BBox = Tuple[float, float, float, float]
+ColorTuple = Tuple[int, int, int, int]
 
 MAX_THREADS = int(os.environ.get("MAX_THREADS", multiprocessing.cpu_count() * 5))