+++ rio_tiler/io/base.py
@@ -7,7 +7,6 @@ import warnings
 from typing import Any, Coroutine, Dict, List, Optional, Sequence, Tuple, Type, Union
 
 import attr
-import numpy
 from morecantile import Tile, TileMatrixSet
 
 from ..constants import WEB_MERCATOR_TMS
@@ -18,6 +17,7 @@ from ..errors import (
     TileOutsideBounds,
 )
 from ..expression import apply_expression
+from ..models import ImageData, ImageStatistics, Info, Metadata, SpatialInfo
 from ..tasks import multi_arrays, multi_values
 
 
@@ -40,14 +40,14 @@ class SpatialMixin:
         )
 
     @property
-    def spatial_info(self) -> Dict:
+    def spatial_info(self) -> SpatialInfo:
         """Return Dataset's spatial info."""
-        return {
-            "bounds": self.bounds,
-            "center": self.center,
-            "minzoom": self.minzoom,
-            "maxzoom": self.maxzoom,
-        }
+        return SpatialInfo(
+            bounds=self.bounds,
+            center=self.center,
+            minzoom=self.minzoom,
+            maxzoom=self.maxzoom,
+        )
 
     def tile_exists(self, tile_z: int, tile_x: int, tile_y: int) -> bool:
         """Check if a tile is inside a the dataset bounds."""
@@ -74,38 +74,37 @@ class BaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractmethod
-    def info(self) -> Dict:
+    def info(self) -> Info:
         """Return Dataset's info."""
         ...
 
     @abc.abstractmethod
-    def stats(self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any) -> Dict:
+    def stats(
+        self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
+    ) -> Dict[str, ImageStatistics]:
         """Return Dataset's statistics."""
         ...
 
-    @abc.abstractmethod
-    def metadata(self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,) -> Dict:
+    def metadata(
+        self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
+    ) -> Metadata:
         """Return Dataset's statistics and info."""
         info = self.info()
-        info["statistics"] = self.stats(pmin, pmax, **kwargs)
-        return info
+        stats = self.stats(pmin, pmax, **kwargs)
+        return Metadata(statistics=stats, **info.dict())
 
     @abc.abstractmethod
-    def tile(
-        self, tile_x: int, tile_y: int, tile_z: int, **kwargs: Any
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    def tile(self, tile_x: int, tile_y: int, tile_z: int, **kwargs: Any) -> ImageData:
         """Read a Map tile from the Dataset."""
         ...
 
     @abc.abstractmethod
-    def part(
-        self, bbox: Tuple[float, float, float, float], **kwargs: Any
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    def part(self, bbox: Tuple[float, float, float, float], **kwargs: Any) -> ImageData:
         """Read a Part of a Dataset."""
         ...
 
     @abc.abstractmethod
-    def preview(self, **kwargs: Any) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    def preview(self, **kwargs: Any) -> ImageData:
         """Return a preview of a Dataset."""
         ...
 
@@ -128,45 +127,42 @@ class AsyncBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractmethod
-    async def info(self) -> Coroutine[Any, Any, Dict]:
+    async def info(self) -> Coroutine[Any, Any, Info]:
         """Return Dataset's info."""
         ...
 
     @abc.abstractmethod
     async def stats(
         self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any
-    ) -> Coroutine[Any, Any, Dict]:
+    ) -> Coroutine[Any, Any, Dict[str, ImageStatistics]]:
         """Return Dataset's statistics."""
         ...
 
     async def metadata(
         self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
-    ) -> Coroutine[Any, Any, Dict]:
+    ) -> Coroutine[Any, Any, Metadata]:
         """Return Dataset's statistics and info."""
         info, stats = await asyncio.gather(
             *[self.info(), self.stats(pmin, pmax, **kwargs)]
         )
-        info["statistics"] = stats
-        return info
+        return Metadata(statistics=stats, **info.dict())
 
     @abc.abstractmethod
     async def tile(
         self, tile_x: int, tile_y: int, tile_z: int, **kwargs: Any
-    ) -> Coroutine[Any, Any, Tuple[numpy.ndarray, numpy.ndarray]]:
+    ) -> Coroutine[Any, Any, ImageData]:
         """Read a Map tile from the Dataset."""
         ...
 
     @abc.abstractmethod
     async def part(
         self, bbox: Tuple[float, float, float, float], **kwargs: Any
-    ) -> Coroutine[Any, Any, Tuple[numpy.ndarray, numpy.ndarray]]:
+    ) -> Coroutine[Any, Any, ImageData]:
         """Read a Part of a Dataset."""
         ...
 
     @abc.abstractmethod
-    async def preview(
-        self, **kwargs: Any
-    ) -> Coroutine[Any, Any, Tuple[numpy.ndarray, numpy.ndarray]]:
+    async def preview(self, **kwargs: Any) -> Coroutine[Any, Any, ImageData]:
         """Return a preview of a Dataset."""
         ...
 
@@ -199,9 +195,9 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
         _re = re.compile(assets.replace("\\\\", "\\"))
         return tuple(set(re.findall(_re, expression)))
 
-    def info(
+    def info(  # type: ignore
         self, assets: Union[Sequence[str], str] = None, *args, **kwargs: Any
-    ) -> Dict:
+    ) -> Dict[str, Info]:
         """Return metadata from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -216,13 +212,13 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
 
         return multi_values(assets, _reader, *args, **kwargs)
 
-    def stats(
+    def stats(  # type: ignore
         self,
         pmin: float = 2.0,
         pmax: float = 98.0,
         assets: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, Dict[str, ImageStatistics]]:
         """Return array statistics from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -237,13 +233,13 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
 
         return multi_values(assets, _reader, pmin, pmax, **kwargs)
 
-    def metadata(
+    def metadata(  # type: ignore
         self,
         pmin: float = 2.0,
         pmax: float = 98.0,
         assets: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, Metadata]:
         """Return metadata from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -269,7 +265,7 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each asset based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Read a Mercator Map tile multiple assets."""
         if not self.tile_exists(tile_z, tile_x, tile_y):
             raise TileOutsideBounds(
@@ -293,14 +289,12 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
                 "assets must be passed either via expression or assets options."
             )
 
-        def _reader(
-            asset: str, *args: Any, **kwargs: Any
-        ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_asset_url(asset)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.tile(*args, **kwargs)
 
-        data, mask = multi_arrays(
+        output = multi_arrays(
             assets,
             _reader,
             tile_x,
@@ -312,9 +306,9 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, assets, data)
+            output.data = apply_expression(blocks, assets, output.data)
 
-        return data, mask
+        return output
 
     def part(
         self,
@@ -325,7 +319,7 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each asset based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Read part of multiple assets."""
         if isinstance(assets, str):
             assets = (assets,)
@@ -344,22 +338,20 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
                 "assets must be passed either via expression or assets options."
             )
 
-        def _reader(
-            asset: str, *args: Any, **kwargs: Any
-        ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_asset_url(asset)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.part(*args, **kwargs)
 
-        data, mask = multi_arrays(
+        output = multi_arrays(
             assets, _reader, bbox, expression=asset_expression, **kwargs,
         )
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, assets, data)
+            output.data = apply_expression(blocks, assets, output.data)
 
-        return data, mask
+        return output
 
     def preview(
         self,
@@ -369,7 +361,7 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each asset based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Return a preview from multiple assets."""
         if isinstance(assets, str):
             assets = (assets,)
@@ -388,20 +380,18 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
                 "assets must be passed either via expression or assets options."
             )
 
-        def _reader(asset: str, **kwargs: Any) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(asset: str, **kwargs: Any) -> ImageData:
             url = self._get_asset_url(asset)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.preview(**kwargs)
 
-        data, mask = multi_arrays(
-            assets, _reader, expression=asset_expression, **kwargs
-        )
+        output = multi_arrays(assets, _reader, expression=asset_expression, **kwargs)
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, assets, data)
+            output.data = apply_expression(blocks, assets, output.data)
 
-        return data, mask
+        return output
 
     def point(
         self,
@@ -472,7 +462,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
 
     def info(
         self, bands: Union[Sequence[str], str] = None, *args, **kwargs: Any
-    ) -> Dict:
+    ) -> Info:
         """Return metadata from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -480,24 +470,30 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         if isinstance(bands, str):
             bands = (bands,)
 
-        def _reader(band: str, **kwargs: Any) -> Dict:
+        def _reader(band: str, **kwargs: Any) -> Info:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.info()
 
         bands_metadata = multi_values(bands, _reader, *args, **kwargs)
-        meta = self.spatial_info
+
+        meta = self.spatial_info.dict()
+
+        # We only keep the value for the first band.
         meta["band_metadata"] = [
-            (ix + 1, bands_metadata[band]["band_metadata"][0][1])
+            (band, bands_metadata[band].band_metadata[0][1])
             for ix, band in enumerate(bands)
         ]
-        meta["band_descriptions"] = [(ix + 1, band) for ix, band in enumerate(bands)]
-        meta["dtype"] = bands_metadata[bands[0]]["dtype"]
-        meta["colorinterp"] = [
-            bands_metadata[band]["colorinterp"][0] for _, band in enumerate(bands)
+        meta["band_descriptions"] = [
+            (band, bands_metadata[band].band_descriptions[0][1])
+            for ix, band in enumerate(bands)
         ]
-        meta["nodata_type"] = bands_metadata[bands[0]]["nodata_type"]
-        return meta
+        meta["dtype"] = bands_metadata[bands[0]].dtype
+        meta["colorinterp"] = [
+            bands_metadata[band].colorinterp[0] for _, band in enumerate(bands)
+        ]
+        meta["nodata_type"] = bands_metadata[bands[0]].nodata_type
+        return Info(**meta)
 
     def stats(
         self,
@@ -505,7 +501,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         pmax: float = 98.0,
         bands: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, ImageStatistics]:
         """Return array statistics from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -516,7 +512,9 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         def _reader(band: str, *args, **kwargs) -> Dict:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
-                return cog.stats(*args, **kwargs)[1]
+                # We only return statistics for Band `1` of each dataset.
+                stats = cog.stats(*args, **kwargs)
+                return stats.get(list(stats)[0])
 
         return multi_values(bands, _reader, pmin, pmax, **kwargs)
 
@@ -526,7 +524,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         pmax: float = 98.0,
         bands: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Metadata:
         """Return metadata from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -534,30 +532,35 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         if isinstance(bands, str):
             bands = (bands,)
 
-        def _reader(band: str, *args, **kwargs) -> Dict:
+        def _reader(band: str, *args, **kwargs) -> Metadata:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
-                meta = cog.metadata(*args, **kwargs)
-                meta["statistics"] = meta["statistics"][1]
-                return meta
+                return cog.metadata(*args, **kwargs)
 
         bands_metadata = multi_values(bands, _reader, pmin, pmax, **kwargs)
 
-        meta = self.spatial_info
+        meta = self.spatial_info.dict()
         meta["band_metadata"] = [
-            (ix + 1, bands_metadata[band]["band_metadata"][0][1])
+            (band, bands_metadata[band].band_metadata[0][1])
             for ix, band in enumerate(bands)
         ]
-        meta["band_descriptions"] = [(ix + 1, band) for ix, band in enumerate(bands)]
-        meta["dtype"] = bands_metadata[bands[0]]["dtype"]
-        meta["colorinterp"] = [
-            bands_metadata[band]["colorinterp"][0] for _, band in enumerate(bands)
+        meta["band_descriptions"] = [
+            (band, bands_metadata[band].band_descriptions[0][1])
+            for ix, band in enumerate(bands)
         ]
-        meta["nodata_type"] = bands_metadata[bands[0]]["nodata_type"]
+        meta["dtype"] = bands_metadata[bands[0]].dtype
+        meta["colorinterp"] = [
+            bands_metadata[band].colorinterp[0] for _, band in enumerate(bands)
+        ]
+        meta["nodata_type"] = bands_metadata[bands[0]].nodata_type
         meta["statistics"] = {
-            band: bands_metadata[band]["statistics"] for _, band in enumerate(bands)
+            # We only keep statistics for Band `1` of each dataset.
+            band: bands_metadata[band].statistics.get(
+                list(bands_metadata[band].statistics)[0]
+            )
+            for _, band in enumerate(bands)
         }
-        return meta
+        return Metadata(**meta)
 
     def tile(
         self,
@@ -570,7 +573,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each band based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Read a Mercator Map tile multiple bands."""
         if not self.tile_exists(tile_z, tile_x, tile_y):
             raise TileOutsideBounds(
@@ -594,14 +597,12 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
                 "bands must be passed either via expression or bands options."
             )
 
-        def _reader(
-            band: str, *args: Any, **kwargs: Any
-        ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(band: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.tile(*args, **kwargs)
 
-        data, mask = multi_arrays(
+        output = multi_arrays(
             bands,
             _reader,
             tile_x,
@@ -613,9 +614,9 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, bands, data)
+            output.data = apply_expression(blocks, bands, output.data)
 
-        return data, mask
+        return output
 
     def part(
         self,
@@ -626,7 +627,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each band based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Read part of multiple bands."""
         if isinstance(bands, str):
             bands = (bands,)
@@ -645,22 +646,20 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
                 "bands must be passed either via expression or bands options."
             )
 
-        def _reader(
-            band: str, *args: Any, **kwargs: Any
-        ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(band: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.part(*args, **kwargs)
 
-        data, mask = multi_arrays(
+        output = multi_arrays(
             bands, _reader, bbox, expression=band_expression, **kwargs,
         )
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, bands, data)
+            output.data = apply_expression(blocks, bands, output.data)
 
-        return data, mask
+        return output
 
     def preview(
         self,
@@ -670,7 +669,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
             str
         ] = "",  # Expression for each band based on index names
         **kwargs: Any,
-    ) -> Tuple[numpy.ndarray, numpy.ndarray]:
+    ) -> ImageData:
         """Return a preview from multiple bands."""
         if isinstance(bands, str):
             bands = (bands,)
@@ -689,18 +688,18 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
                 "bands must be passed either via expression or bands options."
             )
 
-        def _reader(band: str, **kwargs: Any) -> Tuple[numpy.ndarray, numpy.ndarray]:
+        def _reader(band: str, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.preview(**kwargs)
 
-        data, mask = multi_arrays(bands, _reader, expression=band_expression, **kwargs)
+        output = multi_arrays(bands, _reader, expression=band_expression, **kwargs)
 
         if expression:
             blocks = expression.split(",")
-            data = apply_expression(blocks, bands, data)
+            output.data = apply_expression(blocks, bands, output.data)
 
-        return data, mask
+        return output
 
     def point(
         self,