+++ rio_tiler/reader.py
@@ -418,7 +418,7 @@ def stats(
     percentiles: Tuple[float, float] = (2.0, 98.0),
     hist_options: Optional[Dict] = None,
     **kwargs: Any,
-) -> Dict:
+) -> Dict[str, Any]:
     """
     Retrieve statistics from an image.
 
@@ -474,7 +474,7 @@ def stats(
 
     hist_options = hist_options or {}
     return {
-        indexes[b]: raster_stats(data[b], percentiles=percentiles, **hist_options)
+        f"{indexes[b]}": raster_stats(data[b], percentiles=percentiles, **hist_options)
         for b in range(data.shape[0])
     }
 
@@ -551,19 +551,16 @@ def metadata(
 
     hist_options = hist_options or {}
     statistics = {
-        indexes[b]: raster_stats(data[b], percentiles=percentiles, **hist_options)
+        f"{indexes[b]}": raster_stats(data[b], percentiles=percentiles, **hist_options)
         for b in range(data.shape[0])
     }
 
     def _get_descr(ix):
         """Return band description."""
-        name = src_dst.descriptions[ix - 1]
-        if not name:
-            name = "band{}".format(ix)
-        return name
+        return src_dst.descriptions[ix - 1] or ""
 
-    band_descriptions = [(ix, _get_descr(ix)) for ix in indexes]
-    tags = [(ix, src_dst.tags(ix)) for ix in indexes]
+    band_descriptions = [(f"{ix}", _get_descr(ix)) for ix in indexes]
+    tags = [(f"{ix}", src_dst.tags(ix)) for ix in indexes]
 
     other_meta = dict()
     if src_dst.scales[0] and src_dst.offsets[0]: