+++ tests/test_reader.py
@@ -487,10 +487,12 @@ def test_read_unscale():
         numpy.testing.assert_array_equal(mask, maskS)
 
         meta = reader.metadata(src_dst)
-        assert isinstance(meta["statistics"][1]["min"], int)
+        assert isinstance(meta["statistics"]["1"]["min"], int)
+        assert meta["statistics"]["1"]["min"] == -894
 
         meta = reader.metadata(src_dst, unscale=True)
-        assert isinstance(meta["statistics"][1]["min"], float)
+        assert isinstance(meta["statistics"]["1"]["min"], float)
+        assert round(meta["statistics"]["1"]["min"], 1) == 999.9
 
         p = reader.point(src_dst, [310000, 4100000], coord_crs=src_dst.crs)
         assert p == [8917]
@@ -552,7 +554,7 @@ def test_metadata():
         assert meta["colorinterp"] == ["gray"]
         assert meta["scale"] == 0.0001
         assert meta["offset"] == 1000.0
-        assert meta["band_descriptions"] == [(1, "Green")]
+        assert meta["band_descriptions"] == [("1", "Green")]
         assert not meta.get("colormap")
         assert meta["nodata_type"] == "Nodata"